Poprawa argumentu N funkcji search_course

+++ defs.h
@@ -19,5 +19,5 @@ void free_students_grades(struct student_t **students);
 int read_students_grades(const char *filename, struct student_t ***students);
 void display_students_grades(struct student_t **students);
 const struct student_t *search_student(const char *name, const char *last_name, const struct student_t **students);
-const struct course_grades_t *search_course(const char *course, const struct course_grades_t *courses, int N);
+const struct course_grades_t *search_course(const char *course, const struct course_grades_t *courses, unsigned long long N);
 #endif
